[GEOS-9938] Split the Features template module in independent parts. Fix typo in package name
