+++ src/community/features-templating/features-templating-ogcapi/src/main/java/org/geoserver/featurestemplating/ogcapi/GeoJSONAPIWriter.java
@@ -2,7 +2,7 @@
  * This code is licensed under the GPL 2.0 license, available at the root
  * application directory.
  */
-package org.geoserver.featurestemplating.ogapi;
+package org.geoserver.featurestemplating.ogcapi;
 
 import com.fasterxml.jackson.core.JsonGenerator;
 import java.io.IOException;
