+++ src/community/features-templating/features-templating-ogcapi/src/main/java/org/geoserver/featurestemplating/ogcapi/GeoJSONTemplateGetFeatureResponse.java
@@ -2,7 +2,7 @@
  * This code is licensed under the GPL 2.0 license, available at the root
  * application directory.
  */
-package org.geoserver.featurestemplating.ogapi;
+package org.geoserver.featurestemplating.ogcapi;
 
 import static org.geoserver.ogcapi.features.FeatureService.ITEM_ID;
 import static org.springframework.web.context.request.RequestAttributes.SCOPE_REQUEST;
