+++ src/community/features-templating/features-templating-ogcapi/src/main/java/org/geoserver/featurestemplating/ogcapi/JSONTemplateCallBackOGC.java
@@ -2,7 +2,7 @@
  * This code is licensed under the GPL 2.0 license, available at the root
  * application directory.
  */
-package org.geoserver.featurestemplating.ogapi;
+package org.geoserver.featurestemplating.ogcapi;
 
 import java.io.IOException;
 import java.io.OutputStream;
